add timeshift support (closes #3)
The client uses a generic interface for timeshift buffers so that new
ones can be added in the future (e.g. an in-memory buffer). It also
provides a dummy buffer which is used when timeshifting is disabled.
The dummy buffer doesn't really buffer anything, instead it just
passes through all file calls to the underlying input handle.

Since this commit means a different code path is taken when a channel
is started we can now store the currently viewed channel, which makes
implementing signal status (issue #8) possible.